Adapted code for CLion and small code improvements

+++ include/types.h
@@ -15,7 +15,7 @@ namespace hyperspharm
 {
   typedef uint64_t natural_t;
   typedef int64_t integer_t;
-  typedef long double real_t;
+  typedef double real_t;
   typedef std::complex<real_t> complex_t;
   
 }
